refactor/client: get it to compile minimally
