+++ src/lib.rs
@@ -43,6 +43,9 @@
                                    option_unwrap_used))]
 #![cfg_attr(feature="clippy", allow(use_debug, doc_markdown))] // TODO: Fix doc_markdown errors.
 
+// TODO Remove this
+#![allow(unused)]
+
 // extern crate config_file_handler;
 extern crate futures;
 #[macro_use]
@@ -50,14 +53,13 @@ extern crate lazy_static;
 // extern crate libc;
 #[macro_use]
 extern crate log;
-extern crate lru_cache;
 #[macro_use]
 extern crate maidsafe_utilities;
 extern crate rand;
 extern crate routing;
 extern crate rustc_serialize;
 extern crate rust_sodium;
-extern crate self_encryption;
+// extern crate self_encryption;
 // extern crate time;
 extern crate tokio_core;
 #[macro_use]
